[spec/builtin-special] Don't mandate error upon redefinition
Some shells throw a hard error upon redefinition of a special builtin:

- dash
- bash with set -o posix

But this is not mandated.  It's OK to look up a special builtin first,
and allow redefinition.

---

Mark the hard failure of special builtins like 'set' and 'shift' N-I.

bash and busybox don't do it, so I think this behavior will be
surprising, even if it's in POSIX.

---

Also mark bash 'type' confusion as a bug.