[osh] Save shell function source code in a much simpler way!
We now use arena.SnipCodeString() for both shell functions, and blocks,
in the parser.  (Like aliases)

We can delete:

- SaveLinesAndDiscard()
- SnipCodeBlock()

This was supposed to be faster, but gc-cachegrind is slower.

osh-runtime seems OK.

But it IS simpler, and MORE correct.  It fixes the nested
functions/nested blocks code.

It will be easier to optimize something that's both correct and simpler!