Disallow the '--' argument unless `rest` is specified
Follow Python's `argparse` in being strict about the presence of '--'. The
special argument is supported where extraneous positional arguments are
expected and otherwise prohibited.

Add a test to explicitly test that this constraint holds when `--` is the
last positional argument - a situation that seems unlikely to cause harm
and so was previously permitted.