[regtest/bwrap] Got rootless bwrap working
Prior to this change, we used the same strategy as alpine-chroot-install

- bwrap, wrapping 'su -l udu'
  - this starts a login shell, which should read /etc/profile
  - it also sets USER LOGNAME HOME SHELL

Now what we do is:

- Given the user name, use awk to parse (uid, gid, home, shell) out of
  $CHROOT_DIR/etc/passwd
- Then pass flags to bwrap
  * --uid and --gid
  * --clearenv --setenv HOME ...

This works nicely!  We have 11 env vars in the guest shell, just as we
did before.

The next thing would be to start a login shell by invoking with argv0
starting with '-'.  We could use bash 'exec -a', or we could use a
symlink.