[core] Remove libc locale hack; replace with InitLocale()
It sets the locale globally, and doesn't rely on GNU readline to do it.

test/configure-effects.sh has a demo of this.

I added pylib/pylocale.py as a subset of Python-2.7*/Lib/locale.py, so
we don't 'import encodings'.  That isn't available in the oils-ref
tarball.  (This is an annoying complication; maybe rethink it.)

The non-fastlex code path also caused some complication - it could be
rethought.

---

I also add functions to check if the codeset is UTF-8, because Oils
implements UTF-8 internally as well.  We don't want the interpreter to
be inconsistent.

TODO: enable this check, and respects OILS_LOCALE_OK

Revert patch from 2021 commit:

    https://github.com/oils-for-unix/oils/commit/5436affcc33aeaa59d61fefead01667c5c50c85f