[configure] Prefer using mktemp -d (#2386)
Because `mktemp -d` is near universally supported by Unix-likes, use it to
create a random per-execution directory under ${TMPDIR:-/tmp}. This closes
the vector for a well-known attack where pre-known filepaths in /tmp are
modified. Before this change, the caller would need to set TMPDIR manually
to avoid the risk.

---------

Co-authored-by: andychu <andychu@users.noreply.github.com>