[spec/divergence] Add case for nested subshell like ((echo ... (#2399)
gzip has a utility script called zdiff. It gets used at some point in
the check step when building gzip from source. It also makes use of
nested subshells in a few places. The osh parser is tripping over these.
Here's an example from a simple script.

    ((echo foo) | tr o 0)
         ^~~
    _tmp/foo.sh:1: Unexpected token after arithmetic expression (Id.Word_Compound != Id.Arith_RParen)

Other shells seem to handle this case, so in preparation for an inbound
fix, this commit adds a new spec test to cover it.