[ysh] Fix bug in Str.leftMatch() and a|b alternation
We need to respect POSIX ERE precedence by adding ^() around the whole
pattern, not just adding ^ to the left of it.

And then we undo that by ignoring the first 2 indices, which is
match.group(0).

That is, this is incorrect:

    ^[a-z]|[0-9]

because it's parsed as ^[-a-z] OR [0-9].

Instead, leftMatch() now does this:

    ^([a-z]|[0-9])