check argv length before indexing
We have a helper to check if prefix bindings should pe preserved when
running a simple command. This check requires checking the first
argument of the command line, but doesn't check the length before
indexing. This commit patches the helper to return early if argv is
empty. From there, the usual command execution logic will take care of
the rest: either a noop, or erroring out if strict_argv is enabled.