core: Better handle internally-used fds being shifted
Since _SHELL_MIN_FD=100, osh fails to build some Alpine packages:

    $ osh -c 'ulimit -n 64; echo hi >out'
    [ -c flag ]:1: I/O error applying redirect: Invalid argument'

(This is because osh attempts to save stdout's fd before redirection to
a descriptor >=100, which 'ulimit -n 64' doesn't allow)

Currently we can't drop _SHELL_MIN_FD lower, because scripts like this
currently crash osh (users are allowed to use fds<=99):

    $ cat test.osh
    exec 10>out
    echo hello>&10
    cat out
    $ bin/osh test.osh
    hello
    oils I/O error (main): Bad file descriptor

This happens due to osh mishandling "internal" file descriptors, in this
case the script file itself, as can be seen from the following strace
log:

    openat(AT_FDCWD, "test.osh", O_RDONLY)  = 3 <<<<
    fcntl(3, F_DUPFD, 10)                   = 10 <<<<
    close(3)                                = 0
    openat(AT_FDCWD, "out", O_WRONLY|O_CREAT|O_TRUNC, 0666) = 3
    fcntl(10, F_DUPFD, 10)                  = 11 <<<<<
    close(10)                               = 0
    dup2(3, 10)                             = 10
    close(3)                                = 0
    close(11)                               = 0 <<<< it's closed immediately after being saved
    fcntl(1, F_DUPFD, 10)                   = 11
    close(1)                                = 0
    dup2(10, 1)                             = 1
    dup2(11, 1)                             = 1
    close(11)                               = 0
    openat(AT_FDCWD, "out", O_RDONLY)       = 3
    close(3)                                = 0
    hello
    read(10, 0x55bfa557e240, 4096)          = -1 EBADF (Bad file descriptor) <<< still 10
    oils I/O error (main): Bad file descriptor
    close(10)                               = 0

So, add some bookkeeping to FdState - it now tracks whether a file
descriptor is persistent and invokes a callback when such persistent
file descriptors are moved by osh.

Add a callback to FileLineReader - it seek()-s in the new file to
restore the position in the current file and changes file descriptors.

There is currently only one usage of a persistent fd with a callback -
FileLineReader for the script file, since that's what triggers the bug
above. Additional users might need to be changed as more bugs are
discovered - these are currently ignoring the second return value from
fd_state.Open()

Drop _SHELL_MIN_FD to 10, justify the reasoning in a comment.

With these changes, the above test.osh script works as intended, with
internal fds being shifted away from user-requested fds. This
potentially allows us to drop _SHELL_MIN_FD further in the future and
untangle it from the limit on fds users can use in redirects.

Signed-off-by: Andrii Sultanov <sultanovandriy@gmail.com>