[mycpp] Add GC rooting rule and re-enable Souffle (#2449)
This fixes a seg fault in the pretty printer

* mycpp: generate mycpp-soufle build targets

We lost this in the transition to/from souffle being on by default.

* mycpp: fix seg fault in pretty printer

The pretty printer found a gap in the stack roots optimizer. It was
falling over on code with the following form:

    x = alloc_something()...
    do_something(x, function_that_collects())

We werent' generating a root for `x` because its use isn't captured by
the live variable analysis. This commit adds a rule to the rooting
policy to capture this case and adds an example that we can use to check
for regressions.

* change default to souffle
* restore build dep

---------

Co-authored-by: Andy C <andy@oils.pub>