[osh] Make redirects less eager to consume their left-hand side args
Compare bash:

    $ echo x=1> tmp; cat tmp
    x=1

against osh before this fix:

    $ echo x=1> tmp; cat tmp
    x=

Here the redirect absorbs '1' where it shouldn't.

To fix this, "absorb" the preceding argument in WordParser instead of in the
lexer.

Drop the regexes capturing FD_NUM and FD_VAR_NAME as a part of Id.Redir_* in
lexer_def.py. Instead, introduce RedirOp - a new type of word_t which can be
emitted by WordParser when either Lit_Number or Lit_RedirVarName are the only
word in a CompoundWord group before a Redirect interrupts it.

This way, the words leading up to the redirection operator will only be
absorbed when they stand apart from words preceding them (not in 'x=1>',
but in 'x= 1>', for example), given that they follow the rules for
FD_NUM and FD_VAR_NAME.

This also allows us to avoid awkwardly checking the redirection string
for '{' or convertability to an integer, instead reusing the results of
the lexer when matching on Lit_Number|Lit_RedirVarName.

Fixes: #2413

Signed-off-by: Andrii Sultanov <sultanovandriy@gmail.com>