[word_parse] Fix &> and &>> capturing leading descriptors
Previously, these relied on the lack of the regex capturing FD_NUM or
FD_VAR_NAME to avoid capturing a leading descriptor (following bash's
behaviour). Now that every Id.Redir_* dropped the regex and WordParser
joins Id.Lit_Number and Id.Lit_RedirVarName with a Redir, we need to
avoid joining a leading descriptor in word_parser.py for these two types of
redirections.

This makes the redir "3 2&>1 (is it a redirect or is it like a&>1)"
spec test pass again.

Signed-off-by: Andrii Sultanov <sultanovandriy@gmail.com>