[frontend] Fix parsing of say x=1> - redirect should not consume LHS number (#2479)
* [doc] Update outdated link to 'native' in parser-architecture.md
* [spec] Add divergence tests for the 'x=1>' parsing issue
* [frontend] Add Lit_Number and Lit_RedirVarName to the lexer

These shouldn't change anything at runtime, since Lit_* tokens are all
joined into a single word by WordParser, but these will help fix the
x=1> parsing issue.

* [osh] Make redirects less eager to consume their left-hand side args

Compare bash:

    $ echo x=1> tmp; cat tmp
    x=1

against osh before this fix:

    $ echo x=1> tmp; cat tmp
    x=

Here the redirect absorbs '1' where it shouldn't.

To fix this, "absorb" the preceding argument in WordParser instead of in the
lexer.

Drop the regexes capturing FD_NUM and FD_VAR_NAME as a part of Id.Redir_* in
lexer_def.py. Instead, introduce RedirOp - a new type of word_t which can be
emitted by WordParser when either Lit_Number or Lit_RedirVarName are the only
word in a CompoundWord group before a Redirect interrupts it.

This way, the words leading up to the redirection operator will only be
absorbed when they stand apart from words preceding them (not in 'x=1>',
but in 'x= 1>', for example), given that they follow the rules for
FD_NUM and FD_VAR_NAME.

This also allows us to avoid awkwardly checking the redirection string
for '{' or convertability to an integer, instead reusing the results of
the lexer when matching on Lit_Number|Lit_RedirVarName.

Fixes: #2413

* [spec] osh now behaves like Bash in a brace expansion spec test

No need for special casing this anymore

---------

Co-authored-by: Andy C <andy@oils.pub>