[osh] Look ahead to verify '$((' opens an arithmetic substitution
Like cmd_parse with '((', word_parse needs to make sure that '$((' is
closed by uninterrupted parens - '))'. If it is not, the expression is
not an arithmetic substitution - but a command sub with a nested
subshell.

This makes the spec test #8 ("$((which example - command sub versus
arith sub - gnunet-gtk package") in paren-ambiguity pass.

Signed-off-by: Andrii Sultanov <sultanovandriy@gmail.com>