[osh] Look ahead at (( to distinguish arithmetic and commands (#2481)
An expression is arithmetic if:
1) Opening and closing parens are balanced
2) And the closing parentheses are an uninterrupted pair - ))
3) And any other closing parentheses inbetween are closing a
   corresponding nested opening paren

If any of those conditions does not hold, then the expression is instead
treated as two nested subshells.

This is enough to handle the real world use cases (paren-ambiguity spec
tests), but does not follow bash and other shells completely, since
this look-ahead is single-line, biased towards nested subshells and so
can't handle multi-line arithmetic expressions.

Signed-off-by: Andrii Sultanov <sultanovandriy@gmail.com>

This rule can also be used to look ahead for $((, but we haven't enabled that rule yet.

Changes by Andy:

* Make comments more concise
* Put the dparen_hack in one place
* Add unit tests
* Simplify code a bit
  * Don't use the self.line_pos member variable in lookahead - use a local var 'pos'

---------

Signed-off-by: Andrii Sultanov <sultanovandriy@gmail.com>
Co-authored-by: Andy C <andy@oils.pub>