[mycpp] When free function name conflicts with a method, use namespace::func()
In C++, if a method in the current class has the same name as a free function,
it shadows the free function. Whereas 'self.' is required to call a method
in Python, just 'func()' is enough to call a method in C++. This means that
when such a conflicting free function is called in Python, the translated C++
code will call the corresponding method, potentially causing issues.

This came up before, and was only caught accidentally:

https://github.com/oils-for-unix/oils/pull/2241#issuecomment-2629235569

To prevent this from happening, move the const pass earlier, record all the
classes and their methods in MethodDefs (classes namespaces are also stored
in ClassNamespaces). Pass MethodDefs and ClassNamespaces to the impl pass and
if a free function that's being called conflicts with one of the methods of
the current class, call 'current_namespace::function()' instead of just
'function()'. This will make sure the free function and not the method
is called.

This makes the examples/test_func_method_name_conflict.py test pass.

Signed-off-by: Andrii Sultanov <sultanovandriy@gmail.com>