[ysh] Allow $var in $[] for osh
In bash, $[ acts like $((, so for backwards compatibility we need to
accept variable subs like $var inside.

Introduce a new option no_parse_dollar_in_expression_sub that's True by
default in ysh:upgrade, but off in osh, which allows this backwards
compatibility to be maintained.

This makes the new $[$var test in spec/arith pass.

Add a spec test to ysh-expr-sub making sure $[$var is still disallowed
for ysh.

Fixes: #2426

Signed-off-by: Andrii Sultanov <sultanovandriy@gmail.com>