[core] Run builtin cat in a child process (#2534)
Otherwise if it crashes, it brings down the whole subshell with it and
we don't capture its exit status correctly.

Create a new Thunk class for this purpose - its Run() just invokes a builtin
and returns (no exec() involved).

With this change, the divergence test for 'cat' is passing.

Fixes: #2530

Signed-off-by: Andrii Sultanov <sultanovandriy@gmail.com>

Note: builtin cat is enabled by shopt --set rewrite_extern