[mycpp] Fix code generation with methods and free functions of the same name (#2517)
* [mycpp] Add an example of free function and method name conflict

This test currently crashes, since a method is called when a free
function is supposed to be called instead.

* [mycpp] When free function name conflicts with a method, use namespace::func()

In C++, if a method in the current class has the same name as a free function,
it shadows the free function. Whereas 'self.' is required to call a method
in Python, just 'func()' is enough to call a method in C++. This means that
when such a conflicting free function is called in Python, the translated C++
code will call the corresponding method, potentially causing issues.

This came up before, and was only caught accidentally:

https://github.com/oils-for-unix/oils/pull/2241#issuecomment-2629235569

To prevent this from happening, move the const pass earlier, record all the
classes and their methods in MethodDefs (classes namespaces are also stored
in ClassNamespaces). Pass MethodDefs and ClassNamespaces to the impl pass and
if a free function that's being called conflicts with one of the methods of
the current class, call 'current_namespace::function()' instead of just
'function()'. This will make sure the free function and not the method
is called.

This makes the examples/test_func_method_name_conflict.py test pass.

Signed-off-by: Andrii Sultanov <sultanovandriy@gmail.com>

* [mycpp] Update README now that const pass has been moved to an earlier stage

Signed-off-by: Andrii Sultanov <sultanovandriy@gmail.com>

* [mycpp] Split invalid_ctx_raise into two tests

mycpp should detect two errors in invalid_ctx_raise:

1. Class name with __exit__ should start with ctx_ (raised in visitor)
2. No raise in __exit__ (raised in conversion_pass)

But tests showed it was now only detecting the first one.

This used to work since conversion_pass was the first pass - it detected
both errors and bailed out. Now that const_pass has been moved before
conversion_pass, if const_pass sees any errors, conversion_pass will not
be reached (and only error #1 would be reported).

Split these tests in two, so that mycpp bails out after const_pass in
invalid_ctx_name, and after conversion_pass in invalid_ctx_raise.

This way both errors are tested separately.

Signed-off-by: Andrii Sultanov <sultanovandriy@gmail.com>

---------

Signed-off-by: Andrii Sultanov <sultanovandriy@gmail.com>