[osh] Treat unary operator as a word if followed by a binary operator
A construct like "'-a' != 'string'" (when part of a long expression) would
be treated eagerly by the parser: '-a' would be detected as a unary
operator, '!=' as its argument, and 'string' would be an unexpected
trailing word.

Instead, look ahead if the next token after a unary operator is a binary
operator. If so, treat the unary operator as a simple word (it does not
need to be quoted like above, "-a != 'string'" still works correctly)
and continue parsing the binary operator's right hand side argument.

Fixes: #2409

Signed-off-by: Andrii Sultanov <sultanovandriy@gmail.com>