[mycpp] Disallow `break` inside a `switch` (#2554)
* [mycpp] Add invalid-switch-break test

Signed-off-by: Andrii Sultanov <sultanovandriy@gmail.com>

* [mycpp] Disallow break inside a switch

A naive translation (which is currently the case in mycpp) means that a
break inside a switch will have different behavior in C++ compared to
Python (leaving a loop in the latter but only the switch case in the
former). Disallow it to prevent this issue in the future - currently no
code uses a break statement like this.

Fixes: #1482

Signed-off-by: Andrii Sultanov <sultanovandriy@gmail.com>

* [osh] Remove redundant breaks

'done = True' already ensures that the loop will be broken out of. Since
break inside a switch statement behaves differently between C++ and
Python, drop these.

Signed-off-by: Andrii Sultanov <sultanovandriy@gmail.com>

---------

Signed-off-by: Andrii Sultanov <sultanovandriy@gmail.com>