[core] Implement $RANDOM
Create an initial random seed on Mem initilization. Generate a pseudo-random
sequence starting with this seed on every GetValue() of $RANDOM. If
RANDOM is set to any value, treat it as a seed instead. If RANDOM is
unset, $RANDOM will no longer return random values following Bash's
spec.

This makes the builtin-random spec tests pass.

Fixes: #2569

Signed-off-by: Andrii Sultanov <sultanovandriy@gmail.com>