[regtest/aports] Turn on abuild -k -K
To keeps the binary packages and intermediate files

- This makes interactive debugging easier
- And it moves the cleanup responsibility to overlayfs

After each shard, blow away ALL remannts.  Not just the big files.

So the files accumulated by abuild -k -K should be deleted every 100
packages.  I think it wont' fill up the disk.

---

Also start the "shardless" layout with save-package-files.  We're not
relying on this yet, and we still produce the $APORTS_EPOCH/shard{0..15}
output.