[osh] Turn $[] into the synonym of $(( ))
$[] is bash's backwards-compatible synonym of $(( )). OSH used to treat
it as a YSH expression sub instead, hoping that it'd not be used in its
old meaning that much. A few packages from aports are actually using it
as an arithmetic expression, so implement backwards compatibility by
default here.

Add a new parse_ysh_expr_sub option (enabled with ysh:upgrade),
which does turn $[] into a much more powerful YSH expression sub.

Fixes: #2566
Fixes: #2426

Signed-off-by: Andrii Sultanov <sultanovandriy@gmail.com>