[osh/word_parse] Remove unnecessary token mutation
I found this when I started to WRITE something about "pre-structuring
in the lexer".  I used these two examples, of double quoted strings and
here docs:

$ osh --tool tokens -c 'echo "before \" after"'
    0 Lit_Chars            'echo'
    1 WS_Space             ' '
    2 Left_DoubleQuote     '"'
    3 Lit_Chars            'before '
    4 Lit_BackslashDoubleQuote '\\"'
    5 Lit_Chars            ' after'
    6 Right_DoubleQuote    '"'
    7 Eof_Real             ''
(8 tokens)

$ osh --tool tokens -c 'cat <<EOF
> before \" after
> EOF
> '
    0 Lit_Chars            'cat'
    1 WS_Space             ' '
    2 Redir_DLess          '<<'
    3 Lit_Chars            'EOF'
    4 Op_Newline           '\n'
    5 Lit_Chars            'before '
    6 Lit_Chars            '\\"'
    7 Lit_Chars            ' after\n'
    8 Eof_Real             ''
    9 Lit_CharsWithoutPrefix ''
   10 Undefined_Tok        'EOF\n'
   11 Eof_Real             ''
(12 tokens)

I realized that Id.Lit_* are all treated the same by the WordParser and
the evaluator.

So in the second case, we don't need to change the type from
Id.Lit_BackslashDoubleQuote -> Id.Lit_Chars.

I guess this is "hidden invariant" that's useful to document.  This is
why writing is useful!

---

Unrelated:

[devtools] Reverse shortlog