[ASDL cleanup] Remove ast.UserType special case
Prior to this change, we hard-coded the Id_t type into
frontend/syntax.asdl and core/runtime.asdl

Now it is expressed as a 'use' statement in those 2 files.  That means
that the ASDL compiler is INDEPENDENT from Oils code.

However, there is still an issue in C++: we can't "use" non-pointer
types.  So there are still some special cases for 'id_kind_asdl'.

There is also the naming issue.  We use Id.Foo instead of id_e.foo.  We
could make this consistent.